+++ src/main/kotlin/repositories/BaseRepositoryPublisher.kt
@@ -63,7 +63,8 @@ abstract class BaseRepositoryPublisher(
             project.logger.info("Registering repositories")
             registerRepository(this.repositories)
 
-            project.version = versionResolver.getVersion()
+            project.version = versionResolver.getVersionForProject()
+
             if (!isPublishable()) {
                 project.logger.info("Version not publishable, disabling the following tasks:")
                 project.tasks.withType(PublishToMavenRepository::class.java).configureEach {